dev: fixed bitvector length
Former-commit-id: 334f41114551e8bf8cb2178f43f9f2efedaa792c [formerly d6dadb91a59098169ca6115927f0a2aa31c44ba2]
Former-commit-id: 2f3cf59dbc707ab9788797efd22cdbdb55ea1904

+++ src/spexs/reference.go
@@ -38,13 +38,13 @@ func NewReference(size int) *Reference {
 	return ref
 }
 
-func (ref *Reference) Next(idx int, pos int) (Char, int, bool) {
+func (ref *Reference) Next(idx int, pos uint) (Char, uint, bool) {
 	if int(pos) >= len(ref.Seqs[idx].Pat) {
 		return 0, 0, false
 	}
 
-	rune, width := utf8.DecodeRune(ref.Seqs[idx].Pat[pos:])
-	return Char(rune), pos + width, true
+	rune, width := utf8.DecodeRune(ref.Seqs[int(idx)].Pat[pos:])
+	return Char(rune), pos + uint(width), true
 }
 
 func (ref *Reference) ReplaceGroups(pat string) string {
